Mvp2 for main (#16)
* story mode added

* Updated frontend of MVP_1 from Latest_v

* Session Id management frontend

* Set up configuration settings for state management

* streaming and agents as tools to MVP_2

* streaming with type:'loading_message'

* Added UI for Chat History, Control Management

* Updated frontend

* New updated backend

* New changes to mvp_2

* New changes

* RAG Pipeline Implementation

* new changes

* saved, committed

* saved changes

* new changes

* committed changes

* changes committed

* changed chanegs

* saved changes to mvp_2

* committed new changes

* new commits

* committed frontend changes

* 1. Session Management 2. Persistence 3. Feedback Management 4. Backend Optimization 5. Architecture Changes

* removed .env fro git tracking

* 1. Session Management 2. Persistence 3. Feedback Management 4. Backend Optimization 5. Architecture Changes

* Updated requirements.txt

* fixed response error in main.py

* added audio_playback and verse_reader tools

* staged changes

* WIP

* stashed and commited changes before pull

* Update auth flow and backend APIs in MVP_2

* password reset feature and audio playback

* Updated fetures

* Updated fetures

* 1. Added regenerate prompt options\n 2. Added feedback system\n 3. Added report response system\n 4. Session Management

* Voice and Tool Features

* Updated code

* new commits

* Updated code

* Updated code

* stashed changes for local MVP_2 version before pulling remote version

* resolved page.tsx pushed add to git

* Fix UUID handling and image upload functionality

* updated code

* user_id managed in db

* updated code

* updated code

* updated code

* updated code

* session routes

* Saving changes before merge

* saving local changes before merge

* new commits

* Tadabbur MVP Deployement

* set up environments

* fixed the environments

* Fixed story telling agent, functional now

* Most of the build errors fixed

* fixed build issues

* Committing change to page.tsx

* removed page-backup.tsx from frontend

* deployement commits for MVP_2

---------

+++ backend/utils/report_rule.py
@@ -64,12 +64,12 @@ def insert_report_rule(supabase_client, message_id: str, feedback:str, user_id: 
     ) from last_error
 
 
-def delete_report_rule(supabase_client, message_id:str):
+def delete_report_rule(supabase_client, message_id:str, user_id: str):
     # insert rule in the feedback system
     for i in range(8):
         try:
             print("Deleting hard rule...")
-            supabase_client.table('chat_rules').delete().eq("message_id", message_id).execute()
+            supabase_client.table('chat_rules').delete().eq("user_id", user_id).eq("message_id", message_id).execute()
             print(f"✅ Hard rule with message_id {message_id} deleted successfully!")
             return  # Success, return 
         except Exception as e:
